Move ScreenComponent to `/screens` and Add Footer
