+++ src/scss/Footer.scss
@@ -1,0 +1,6 @@
+.footer {
+  position: fixed;
+  bottom: 0;
+  width: 100%;
+  height: 50px;
+}
